Separate out the different kernel packages
