+++ src/kernel/core/char.c
