+++ src/kernel/core/char.h
