+++ src/kernel/core/core.h
@@ -1,0 +1,17 @@
+// Declarations of values defined in Elm code
+// that are referenced from C code
+#ifndef CORE_H
+#define CORE_H
+
+#include "../wrapper/wrapper.h"
+#include "basics.h"
+#include "char.h"
+#include "debug.h"
+#include "elm.h"
+#include "gc.h"
+#include "list.h"
+#include "string.h"
+#include "types.h"
+#include "utils.h"
+
+#endif
