+++ src/kernel/core/debug.c
