+++ src/kernel/core/gc-internals.c
