+++ src/kernel/core/gc-internals.h
