+++ src/kernel/core/gc.c
