+++ src/kernel/core/gc.h
