+++ src/kernel/core/list.c
