+++ src/kernel/core/list.h
