+++ src/kernel/core/string.c
