+++ src/kernel/core/string.h
