+++ src/kernel/core/utils.c
