+++ src/kernel/core/utils.h
