+++ src/kernel/wrapper/wrapper.h
@@ -1,7 +1,7 @@
 #ifndef WRAPPER_H
 #define WRAPPER_H
 
-#include "../types.h"
+#include "../core/types.h"
 
 // Compiler-generated app-specific data
 extern FieldGroup* Wrapper_appFieldGroups[];
