+++ src/test/gc_test.c
@@ -1,11 +1,9 @@
-#include "../kernel/gc.h"
-
 #include <stdbool.h>
 #include <stdlib.h>
 #include <string.h>
 
-#include "../kernel/gc-internals.h"
-#include "../kernel/kernel.h"
+#include "../kernel/core/core.h"
+#include "../kernel/core/gc-internals.h"
 #include "gc/replay_test.h"
 #include "gc/stackmap_mark_test.h"
 #include "test.h"
