+++ src/test/test.h
@@ -4,7 +4,7 @@
 #define __USE_MINGW_ANSI_STDIO 1
 #include <stdio.h>
 
-#include "../kernel/types.h"
+#include "../kernel/core/types.h"
 
 enum {
   CTOR_Nothing,
