+++ src/test/types_test.c
@@ -1,4 +1,4 @@
-#include "../kernel/types.h"
+#include "../kernel/core/types.h"
 #include <string.h>
 #include <unistd.h>
 #include "test.h"
